+++ buildSrc/src/main/kotlin/Configurations.kt
@@ -47,7 +47,6 @@ object Versions {
 
     object Kotlin {
         const val core = "1.3.61"
-        const val coroutines = "1.3.3"
     }
 
     object AndroidX {
@@ -69,9 +68,7 @@ object Dependencies {
     object Kotlin {
         const val classpath = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.Kotlin.core}"
         const val stdlib = "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.Kotlin.core}"
-        const val coroutinesAndroid = "org.jetbrains.kotlinx:kotlinx-coroutines-android:${Versions.Kotlin.coroutines}"
         const val test = "org.jetbrains.kotlin:kotlin-test-junit:${Versions.Kotlin.core}"
-        const val coroutinesTest = "org.jetbrains.kotlinx:kotlinx-coroutines-test:${Versions.Kotlin.coroutines}"
     }
 
     object AndroidX {